Add tests for UUID generator
